application/requirements.txt
grpcio>=1.53.2
# grpcio-tools==1.36.1  # still needed?
protobuf<4.0dev
pyuwsgi
Django
pyyaml
requests
randomname
tqdm
mysql
mysqlclient
django-webpack-loader==0.6.0
python-gitlab
GitPython
python-slugify
git+https://github.com/allenai/ir_datasets
git+https://github.com/mam10eks/diffir
pandas
markdown
PyGithub==1.59.1
ghapi
parameterized
mockito
approvaltests==7.3.0
django-extensions
coverage
coverage-badge
discourse-client-in-disraptor==0.0.8
tira>=0.0.97
huggingface-hub
djangorestframework==3.15.1
djangorestframework-jsonapi['django-filter']==7.0.0